+++ CHANGELOG.md
@@ -12,6 +12,14 @@
 - [#4752](https://github.com/influxdata/telegraf/pull/4752): Add per-directory file counts in the filecount input.
 - [#4811](https://github.com/influxdata/telegraf/pull/4811): Add windows service name lookup to procstat input.
 - [#4807](https://github.com/influxdata/telegraf/pull/4807): Add entity-body compression to http output.
+- [#4838](https://github.com/influxdata/telegraf/pull/4838): Add telegraf version to User-Agent header.
+
+## v1.8.2 [unreleased]
+
+### Bugfixes
+
+- [#4844](https://github.com/influxdata/telegraf/pull/4844): Update write path to match updated InfluxDB v2 API.
+- [#4840](https://github.com/influxdata/telegraf/pull/4840): Fix missing timeouts in vsphere input.
 
 ## v1.8.1 [2018-10-03]